PB-903 Rename "translation" to more general "language"
